requirements.txt
requests
PyGithub
gql==3.0.0a5
python-dotenv